Aded Network Controller. Worked on Sniper.
Have added a network controller for PUN2, it is working offline, for testing purposes, but it is allowes for Photon Network Calls to function without errors. Worked on Sniper functionality. Added a lot of Instrumentation DBG, to test the code. And A Plugin I use to debug and Analyse code called ConsoleE Free version.