Sniper joystick reset when disabling.
Some adjusting of Sniping mechanics. And instrumentation changes. Removed the Error in the unused snipershooting script.

Edited the gitignore to exclude ProjectSettings.asset and PhotonServerSettings.asset ,PhotonServerSettings.asset.meta. This is because we use a different keystore in  ProjectSettings.asset, and using version controlling PhotonServerSettings.asset  causes unity to get stuck on import.  They should be local individual configurations , not shared.