Sniper v3.1
Modified Sniper further to meet specific reuirements.
Sniper works by communicating with an animator.

Added a plugin, for debugging, shows hierarchy and inspector while running the apk in the phone. Must be added to a scene canvas for it to run. Not currently running. Used it to find a second camera affecting framerate.